requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-multipart
pillow
python-dotenv
requests
openai>=1.13.3
aiohttp
ffmpeg-python
PyJWT
pytesseract
psycopg[binary]>=3.3.2
